+++ FloconDesktop/data/local/src/commonMain/kotlin/io/github/openflocon/data/local/network/models/badquality/BadQualityConfigEntity.kt
@@ -14,7 +14,7 @@ data class BadQualityConfigEntity(
 )
 
 data class LatencyConfigEmbedded(
-    val triggerProbability: Float,
+    val triggerProbability: Double,
     val minLatencyMs: Long,
     val maxLatencyMs: Long,
 )
